refactor(actions): make Actions component uniform and reusable
fix #191
