+++ src/internal/Actions/Action.module.css
@@ -1,0 +1,84 @@
+@import url("@/lib/Layout/customMedia.css");
+
+.action-button {
+    --icon-size: calc(var(--fg-size-unit) * 4);
+
+    align-items: center;
+    border-top: 1px solid var(--fg-background-100);
+    box-sizing: border-box;
+    cursor: pointer;
+    display: flex;
+    flex-direction: row;
+    flex-grow: 1;
+    flex-shrink: 1;
+    font-weight: var(--fg-font-weight-bolder);
+    justify-content: center;
+    overflow: hidden;
+    padding: calc(var(--fg-size-unit) * 2);
+    -webkit-tap-highlight-color: transparent;
+    text-overflow: ellipsis;
+    transition:
+        background-color var(--fg-time-sm) ease-in-out,
+        color var(--fg-time-sm) ease-in-out;
+    user-select: none;
+    white-space: nowrap;
+
+    &:hover {
+        background: var(--fg-color-do);
+        color: var(--fg-background-000);
+    }
+
+    &:focus-visible:not(:hover) {
+        color: var(--fg-color-do);
+    }
+
+    &:active {
+        background: var(--fg-color-do);
+        color: var(--fg-background-000);
+
+        & .action-label,
+        & .action-icon {
+            left: 3px;
+            position: relative;
+            top: 3px;
+        }
+    }
+
+    @media (--viewport-md) {
+        &:not(:only-child, :last-child) {
+            border-right: 1px solid var(--fg-background-100);
+        }
+
+        &:not(:only-child) {
+            width: 50%;
+        }
+    }
+}
+
+.default {
+    color: var(--fg-color-sol);
+}
+
+.link {
+    color: var(--fg-color-action);
+}
+
+.success {
+    color: var(--fg-color-success);
+}
+
+.danger {
+    color: var(--fg-color-error);
+}
+
+.action-icon {
+    flex-shrink: 0;
+    margin-right: var(--fg-size-unit);
+    width: var(--icon-size);
+}
+
+.action-label {
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+}
