+++ src/internal/Actions/index.ts
@@ -1,1 +1,2 @@
 export {ActionTypes} from './ActionTypes.ts';
+export {Action} from './Action.tsx';
