+++ src/lib/InputCheckbox/InputCheckbox.module.css
@@ -27,9 +27,11 @@
     width: 24px;
 
     &::before {
+        --size: calc((var(--fg-size-unit) * 2) - 2px);
+
         background-color: var(--center-color);
         content: "";
-        height: calc(var(--fg-size-unit) * 1.5);
+        height: var(--size);
         left: 50%;
         opacity: 0;
         position: absolute;
@@ -38,7 +40,7 @@
         transition:
             background-color var(--fg-time-sm) ease-in-out,
             opacity var(--fg-time-sm) ease-in-out;
-        width: calc(var(--fg-size-unit) * 1.5);
+        width: var(--size);
     }
 
     &:--disabledSelector {
