+++ src/lib/Toast/Toast.module.css
@@ -118,80 +118,10 @@
     }
 }
 
-.action-button {
-    align-items: center;
-    border-top: 1px solid var(--fg-background-100);
-    box-sizing: border-box;
-    cursor: pointer;
-    display: flex;
-    flex-direction: row;
-    flex-grow: 1;
-    flex-shrink: 1;
-    font-size: var(--fg-font-size-extra-small);
+.toast .action-button {
+    --icon-size: calc(var(--fg-size-unit) * 3);
+
+    font-size: var(--fg-font-size-small);
     font-weight: var(--fg-font-weight-bolder);
-    justify-content: center;
-    line-height: 24px;
-    overflow: hidden;
     padding: var(--fg-size-unit);
-    -webkit-tap-highlight-color: transparent;
-    text-overflow: ellipsis;
-    transition:
-        background-color var(--fg-time-sm) ease-in-out,
-        color var(--fg-time-sm) ease-in-out;
-    user-select: none;
-    white-space: nowrap;
-
-    &:hover {
-        background: var(--fg-color-do);
-        color: var(--fg-background-000);
-    }
-
-    &:focus-visible:not(:hover) {
-        color: var(--fg-color-do);
-    }
-
-    &:active {
-        & .action-label,
-        & .action-icon {
-            left: 2px;
-            position: relative;
-            top: 2px;
-        }
-    }
-
-    &:not(:only-child, :last-child) {
-        border-right: 1px solid var(--fg-background-100);
-    }
-
-    &:not(:only-child) {
-        width: 50%;
-    }
-}
-
-.default {
-    color: var(--fg-color-sol);
-}
-
-.link {
-    color: var(--fg-color-action);
-}
-
-.success {
-    color: var(--fg-color-success);
-}
-
-.danger {
-    color: var(--fg-color-error);
-}
-
-.action-icon {
-    flex-shrink: 0;
-    margin-right: calc(var(--fg-size-unit) / 2);
-    width: calc(var(--fg-size-unit) * 3);
-}
-
-.action-label {
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
 }
